% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_power.R
\name{mod_power_ui}
\alias{mod_power_ui}
\alias{mod_power_server}
\title{mod_power_ui and mod_power_server}
\usage{
mod_power_ui(id)

mod_power_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
