% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_probit_regression.R
\name{mod_probit_regression_ui}
\alias{mod_probit_regression_ui}
\alias{mod_probit_regression_server}
\title{mod_probit_regression_ui and mod_probit_regression_server}
\usage{
mod_probit_regression_ui(id)

mod_probit_regression_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
